fix osb

+++ proys/despachos/back/app/Controllers/Api/V1/ClientesController.php
@@ -72,6 +72,15 @@ class ClientesController extends ResourceController
      */
     public function create()
     {
+        header('Access-Control-Allow-Origin: *');
+
+        header("Access-Control-Allow-Headers: X-API-KEY, Origin, X-Requested-With, Content-Type, Accept, Access-Control-Request-Method, Authorization");
+        header("Access-Control-Allow-Methods: GET, POST, OPTIONS, PUT, DELETE, PATCH");
+        header("Allow: GET, POST, OPTIONS, PUT, DELETE, PATCH");
+        $method = $_SERVER['REQUEST_METHOD'];
+        if ($method == "OPTIONS") {
+            die();
+        }        
         $data = $this->request->getJSON();
 
         $data->created_at = $this->datetimeNow->format('Y-m-d H:i:s');
@@ -101,6 +110,15 @@ class ClientesController extends ResourceController
      */
     public function update($id = null)
     {
+        header('Access-Control-Allow-Origin: *');
+
+        header("Access-Control-Allow-Headers: X-API-KEY, Origin, X-Requested-With, Content-Type, Accept, Access-Control-Request-Method, Authorization");
+        header("Access-Control-Allow-Methods: GET, POST, OPTIONS, PUT, DELETE, PATCH");
+        header("Allow: GET, POST, OPTIONS, PUT, DELETE, PATCH");
+        $method = $_SERVER['REQUEST_METHOD'];
+        if ($method == "OPTIONS") {
+            die();
+        }
         $validateEntry = $this->model->find($id);
         if (empty($validateEntry)) {
             return $this->failNotFound();
@@ -135,6 +153,15 @@ class ClientesController extends ResourceController
     public function delete($id = null)
     {
 
+        header('Access-Control-Allow-Origin: *');
+
+        header("Access-Control-Allow-Headers: X-API-KEY, Origin, X-Requested-With, Content-Type, Accept, Access-Control-Request-Method, Authorization");
+        header("Access-Control-Allow-Methods: GET, POST, OPTIONS, PUT, DELETE, PATCH");
+        header("Allow: GET, POST, OPTIONS, PUT, DELETE, PATCH");
+        $method = $_SERVER['REQUEST_METHOD'];
+        if ($method == "OPTIONS") {
+            die();
+        }
         if ($this->model->delete($id)) {
             return $this->respondDeleted($id);
         } else {
